Clean up HomeView chart section layout
